requirements.txt
# web
mangum
quart

# database
aiomysql
pymongo

# dotenv
python-dotenv
